Begin work on amended replay format
