+++ YARG.Core/Replays/Analyzer/ReplayAnalyzer.cs
@@ -16,7 +16,7 @@ namespace YARG.Core.Replays.Analyzer
     public class ReplayAnalyzer
     {
         private readonly SongChart _chart;
-        private readonly Replay    _replay;
+        private readonly ReplayNew    _replay;
 
         private readonly double _fps;
         private readonly bool   _doFrameUpdates;
@@ -25,7 +25,7 @@ namespace YARG.Core.Replays.Analyzer
 
         private readonly Random _random = new();
 
-        public ReplayAnalyzer(SongChart chart, Replay replay, double fps, bool keepEngineLoggers)
+        public ReplayAnalyzer(SongChart chart, ReplayNew replay, double fps, bool keepEngineLoggers)
         {
             _chart = chart;
             _replay = replay;
@@ -37,7 +37,7 @@ namespace YARG.Core.Replays.Analyzer
         }
 
         public static AnalysisResult[] AnalyzeReplay(SongChart chart,
-            Replay replay, double fps = 0, bool keepEngineLoggers = false)
+            ReplayNew replay, double fps = 0, bool keepEngineLoggers = false)
         {
             var analyzer = new ReplayAnalyzer(chart, replay, fps, keepEngineLoggers);
             return analyzer.Analyze();
